requirements.txt
letta>=0.6.0
letta-client>=0.1.0
asyncpg>=0.29.0
pgvector>=0.2.0
readability-lxml>=0.8.0
beautifulsoup4>=4.12.0
requests>=2.31.0
click>=8.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
anthropic>=0.39.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pypdf>=3.17.0
